fix: Mouse/canvas misalignment bug
The mouse position matched the drawing coordinate at the top left corner
of the canvas, but the farther from the top left it was, the more
"drift" it would accumulate - eventually the game would think the mouse
was hovering a vertex that was visually about 16 pixels to the left and
about 10 pixels above. Turned out it was due to not setting overflow:
hidden on body, causing the disabled scrollbar to shrink the SVG element
and thus make a bit smaller than the window size.